requirements.txt

streamlit>=1.36
yfinance>=0.2.40
pandas>=2.2
numpy>=1.26
python-dateutil>=2.9
